+++ src/features/letter-junior-list/index.tsx
@@ -1,1 +1,2 @@
 export { default as JuniorLetterRecentListFeature } from './containers/LetterRecentListContainer'
+export { default as JuniorLetterArchiveListFeature } from './containers/LetterArchiveListContainer'
