+++ src/pages/latte-chat/letter/archive/junior/ui.tsx
@@ -1,0 +1,5 @@
+import { JuniorLetterArchiveListFeature } from '@/features/letter-junior-list'
+
+export default function JuniorLetterArchiveListPage() {
+  return <JuniorLetterArchiveListFeature />
+}
